@@ -2,7 +2,7 @@
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
+    <TargetFramework>net6.0</TargetFramework>
     <LangVersion>latest</LangVersion>
     <Configurations>Debug;Release</Configurations>
   </PropertyGroup>
@@ -10,14 +10,14 @@
   <Choose>
     <When Condition="'$(Configuration)'=='Release' Or '$(Configuration)'=='Debug'">
       <ItemGroup>
-        <PackageReference Include="Discord.Net.Labs.Commands" Version="2.3.5" />
+        <PackageReference Include="Discord.Net.Labs.Commands" Version="3.6.1" />
       </ItemGroup>
     </When>
   </Choose>
 
   <ItemGroup>
-    <PackageReference Include="GScraper" Version="1.0.1" />
-    <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="5.0.2" />
+    <PackageReference Include="GScraper" Version="1.0.2" />
+    <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="6.0.0" />
     <ProjectReference Include="..\src\Fergun.Interactive.csproj" />
   </ItemGroup>
 
