@@ -8,44 +8,45 @@ using ExampleBot.Services;
 using Fergun.Interactive;
 using Microsoft.Extensions.DependencyInjection;
 
-namespace ExampleBot
+namespace ExampleBot;
+
+internal static class Program
 {
-    internal class Program
+    private static Task Main() => StartAsync();
+
+    private static async Task StartAsync()
     {
-        private static Task Main() => new Program().StartAsync();
+        var services = ConfigureServices();
+        var client = services.GetRequiredService<DiscordSocketClient>();
 
-        private async Task StartAsync()
-        {
-            var services = ConfigureServices();
-            var client = services.GetRequiredService<DiscordSocketClient>();
+        services.GetRequiredService<CommandService>().Log += LogAsync;
+        services.GetRequiredService<InteractiveService>().Log += LogAsync;
+        client.Log += LogAsync;
 
-            services.GetRequiredService<CommandService>().Log += LogAsync;
-            client.Log += LogAsync;
+        // Put your token here
+        string token = "Token";
+        await client.LoginAsync(TokenType.Bot, token);
+        await client.StartAsync();
 
-            // Put your token here
-            string token = "Token";
-            await client.LoginAsync(TokenType.Bot, token);
-            await client.StartAsync();
+        await services.GetRequiredService<CommandHandlingService>().InitializeAsync();
 
-            await services.GetRequiredService<CommandHandlingService>().InitializeAsync();
-
-            await Task.Delay(Timeout.Infinite);
-        }
-
-        private Task LogAsync(LogMessage log)
-        {
-            Console.WriteLine(log.ToString());
-
-            return Task.CompletedTask;
-        }
-
-        private IServiceProvider ConfigureServices()
-            => new ServiceCollection()
-                .AddSingleton<DiscordSocketClient>()
-                .AddSingleton<CommandService>()
-                .AddSingleton<CommandHandlingService>()
-                .AddSingleton<InteractiveService>()
-                .AddSingleton<Random>()
-                .BuildServiceProvider();
+        await Task.Delay(Timeout.Infinite);
     }
+
+    private static Task LogAsync(LogMessage log)
+    {
+        Console.WriteLine(log.ToString());
+
+        return Task.CompletedTask;
+    }
+
+    private static IServiceProvider ConfigureServices()
+        => new ServiceCollection()
+            .AddSingleton(new DiscordSocketConfig { LogLevel = LogSeverity.Verbose })
+            .AddSingleton(new CommandServiceConfig { LogLevel = LogSeverity.Verbose })
+            .AddSingleton<DiscordSocketClient>()
+            .AddSingleton<CommandService>()
+            .AddSingleton<CommandHandlingService>()
+            .AddSingleton<InteractiveService>()
+            .BuildServiceProvider();
 }
